🔧 chore: merge 충돌 해결

+++ backend/JiShop/src/main/java/com/jishop/popular/service/PopularServiceImpl.java
@@ -107,7 +107,7 @@ public class PopularServiceImpl implements PopularService {
                 product.getName(),
                 product.getOriginPrice(),
                 product.getDiscountPrice(),
-                productScore.getProduct().getDiscountRate().intValue(),
+                productScore.getProduct().getDiscountRate(),
                 productScore.getTotalOrderCount(),
                 productScore.getReviewRating().intValue()
         );
